+++ app/src/main/kotlin/org/breezyweather/sources/msn/json/MsnHourly.kt
@@ -1,0 +1,42 @@
+/*
+ * This file is part of Breezy Weather.
+ *
+ * Breezy Weather is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, version 3 of the License.
+ *
+ * Breezy Weather is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+ * License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Breezy Weather. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.breezyweather.sources.msn.json
+
+import kotlinx.serialization.Serializable
+import org.breezyweather.common.serializer.DateSerializer
+import java.util.Date
+
+@Serializable
+data class MsnHourly(
+    @Serializable(DateSerializer::class) val valid: Date?,
+    val cap: String?,
+    val symbol: String?,
+    val temp: Double?,
+    val feels: Double?,
+    val rh: Double?,
+    val dewPt: Double?,
+    val baro: Double?,
+    val vis: Double?,
+    val cloudCover: Double?,
+    val uv: Double?,
+    val windDir: Double?,
+    val windSpd: Double?,
+    val windGust: Double?,
+    val precip: Double?,
+    val rainAmount: Double?,
+    val snowAmount: Double?,
+)